make `CallInfo` propagate effects
Add `call_effects(::CallInfo) -> Effects` feature to the `CallInfo`
interface. This change will make it easier for the inlining algorithm to
see the effects of a call in question.

This commit also setups new `ModifyFieldInfo` callinfo type so that it
can propagate `modifyfield!`-specific information to the inliner, which
implements a special handling for the call.